refactor: update interface types
